requirements.txt
pytest
python-dotenv
Faker
requests
PyHamcrest
PyYAML
allure-pytest